fix(consonant-blocks): clean up load-header
